Extract stateless functions createIdToken, signJwtPresentation, generateOpenid4vciProof
